Add Virtual DRM device driver
Virtual DRM splits the resources of an overlay plane into multiple
virtual devices to allow each plane to be accessed by each process.

This makes it possible to overlay images output from multiple processes
on a display. For example, one process displays the camera image without
compositor while another process overlays the compositor's drawing of
the UI.

The virtual DRM creates standalone virtual device and make DRM planes
from a master device (e.g. card0) accessible via one or more virtual
devices. However, these plane are no longer accessible from the original
device.
Each virtual device (and plane) can be accessed via a separate
device file.

It is necessary that a virtual DRM device node is created in the device
tree. A plane node is created as a child node of the vDRM device node,
and the initial position (x, y), size (width, height) and z-position of
the plane shall be described in the plane node. And, "virt-drm" is
specified for compatible of the vDRM device node.

For example:
----------
&soc {
    vdrm0: vdrm@0 {
        compatible = "virt-drm";
        plane@0 {
            x = <100>;
            y = <100>;
            width = <800>;
            height = <600>;
            zpos = <1>;
        };
    };
};
----------